season can't deleted if it associated with

+++ app/Http/Controllers/SeasonController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\SeasonRequest;
 use App\Http\Requests\UpdateSeasonRequest;
 use Illuminate\Support\Carbon;
 use Illuminate\Validation\ValidationException;
+use Illuminate\Support\Facades\DB;
 
 use App\Season;
 use App\Supplier;
@@ -136,7 +137,18 @@ class SeasonController extends Controller
      */
     public function destroy($id)
     {
-        Season::destroy(decrypt($id));
-        return redirect()->route('seasons.index')->with('success_message', 'Season deleted successfully');
+        $quotes                      = DB::table('quotes')->where('season_id', decrypt($id))->count();
+        $bookings                    = DB::table('bookings')->where('season_id', decrypt($id))->count();
+        $commission_criteria_seasons = DB::table('commission_criteria_seasons')->where('season_id', decrypt($id))->count();
+        $supplier_bulk_payments      = DB::table('supplier_bulk_payments')->where('season_id', decrypt($id))->count();
+        $supplier_rate_sheets        = DB::table('supplier_rate_sheets')->where('season_id', decrypt($id))->count();
+        $templates                   = DB::table('templates')->where('season_id', decrypt($id))->count();
+
+        if($quotes == 0 && $bookings == 0 && $commission_criteria_seasons == 0 && $supplier_bulk_payments == 0 && $supplier_rate_sheets == 0 && $templates == 0){
+            Season::destroy(decrypt($id));
+            return redirect()->route('seasons.index')->with('success_message', 'Season deleted successfully');
+        }
+
+        return redirect()->route('seasons.index')->with('success_message', "Season can not deleted beacuse it is associated one or more record.");
     }
 }
